split reducers into multiple files
